# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/post/list": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "List posts",
                "description": "Endpoint to retrieve a list of posts.",
                "operationId": "listPosts",
                "responses": {
                    "200": {
                        "description": "List of posts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID of the post",
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "caption": {
                                                "description": "Caption of the post",
                                                "type": "string",
                                                "example": "This is a post caption"
                                            },
                                            "user_id": {
                                                "description": "ID of the user who created the post",
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "description": "Timestamp when the post was created",
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-01-01T00:00:00.000Z"
                                            },
                                            "updated_at": {
                                                "description": "Timestamp when the post was updated",
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-01-01T00:00:00.000Z"
                                            },
                                            "images": {
                                                "description": "Images associated with the post",
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "image_url": {
                                                            "description": "URL of the image",
                                                            "type": "string",
                                                            "example": "http://example.com/image.jpg"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "videos": {
                                                "description": "Videos associated with the post",
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "video_url": {
                                                            "description": "URL of the video",
                                                            "type": "string",
                                                            "example": "http://example.com/video.mp4"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/post/create": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Create a new post",
                "description": "Endpoint to create a new post.",
                "operationId": "createPost",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "caption": {
                                        "description": "Caption of the post",
                                        "type": "string",
                                        "example": "New post caption"
                                    },
                                    "image": {
                                        "type": "array",
                                        "items": {
                                            "description": "URL of an image associated with the post",
                                            "type": "string",
                                            "example": "http://example.com/image.jpg"
                                        }
                                    },
                                    "video": {
                                        "type": "array",
                                        "items": {
                                            "description": "URL of a video associated with the post",
                                            "type": "string",
                                            "example": "http://example.com/video.mp4"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID of the created post",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "caption": {
                                            "description": "Caption of the post",
                                            "type": "string",
                                            "example": "New post caption"
                                        },
                                        "user_id": {
                                            "description": "ID of the user who created the post",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "description": "Timestamp when the post was created",
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-01-01T00:00:00.000Z"
                                        },
                                        "updated_at": {
                                            "description": "Timestamp when the post was updated",
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-01-01T00:00:00.000Z"
                                        },
                                        "images": {
                                            "description": "Images associated with the post",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "image_url": {
                                                        "description": "URL of the image",
                                                        "type": "string",
                                                        "example": "http://example.com/image.jpg"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "videos": {
                                            "description": "Videos associated with the post",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "video_url": {
                                                        "description": "URL of the video",
                                                        "type": "string",
                                                        "example": "http://example.com/video.mp4"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid input data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}